+++ StreetLighting/Models/RespondentName.cs
@@ -5,6 +5,7 @@ namespace StreetLighting.Models
     public class RespondentName
     {
         [Required]
+        [StringLength(50, ErrorMessage = "The {0} value cannot exceed {1} characters. ")]
         public string FullName { get; set; }
     }
 }
