+++ StreetLighting/Views/Shared/_Layout.cshtml
@@ -74,7 +74,8 @@
     </footer>
 
     <script src="~/lib/jquery/dist/jquery.min.js"></script>
-    @*<script src="~/lib/bootstrap/dist/js/bootstrap.bundle.min.js"></script>*@
+    @*<script src="~/lib/jquery-validation/dist/jquery.validate.min.js"></script>
+    <script src="~/lib/jquery-validation-unobtrusive/jquery.validate.unobtrusive.min.js"></script>*@
     <script src="~/js/site.js" asp-append-version="true"></script>
     <script src="~/js/govuk-frontend-3.6.0.min.js" asp-append-version="true"></script>
     <script>window.GOVUKFrontend.initAll()</script>
